test: simplify esm test
This only maintains testing for esm, not the typescript support.
It also does not make any changes to the existing CircleCI setup.

With self-referential modules we can simply "require" or "import"
a module from within itself. This is also a decent way to sniff
for support for ESM, as all modern versions of Node.js that support
ESM have this functionality.

Feel free to let me know if you want me to remove the `test-import`
directory or if you want to update circleCI.